+++ BambooTracker/gui/configuration_dialog.cpp
@@ -105,3 +105,38 @@ void ConfigurationDialog::on_ConfigurationDialog_accepted()
 	config_.lock()->setSampleRate(ui->sampleRateComboBox->currentData(Qt::UserRole).toInt());
 	config_.lock()->setBufferLength(ui->bufferLengthHorizontalSlider->value());
 }
+
+void ConfigurationDialog::on_generalSettingsListWidget_itemSelectionChanged()
+{
+	QString text;
+	switch (ui->generalSettingsListWidget->currentRow()) {
+	case 0:	// Warp cursor
+		text = "Warp the cursor around the edges of the pattern editor.";
+		break;
+	case 1:	// Warp across orders
+		text = "Move to previous or next order when reaching top or bottom in the pattern editor.";
+		break;
+	case 2:	// Show row numbers in hex
+		text = "Display order numbers and the order count on the status bar in hexadecimal.";
+		break;
+	case 3:	// Preview previous/next orders
+		text = "Preview previous and next orders in the pattern editor.";
+		break;
+	case 4:	// Backup modeles
+		text = "Create a backup copy of the existing file when saving a module.";
+		break;
+	case 5:	// Don't select on double click
+		text = "Don't select the whole track when double-clicking in the pattern editor.";
+		break;
+	case 6:	// Reverse FM volume order
+		text = "Reverse the order of FM volume so that 00 is the quietest in the pattern editor";
+		break;
+	case 7:	// Move cursor to right
+		text = "Move the cursor to right after entering effects in the pattern editor.";
+		break;
+	default:
+		text = "";
+		break;
+	}
+	ui->descPlainTextEdit->setPlainText(QString("Description: ") + text);
+}